cambios de iconos y redireccion de ventanas
