+++ src/components/appointment/AppointmentList.tsx
@@ -1,0 +1,127 @@
+"use client";
+
+import { useEffect, useState } from "react";
+import { Eye } from "lucide-react";
+import { Appointment } from "@/lib/appointment/IAppointment";
+import { getAppointmentByPetId } from "@/lib/appointment/getAppointmentByPetId";
+import { formatDate } from "@/lib/utils";
+import GenericTable, { Column, PaginationInfo, TableAction } from "@/components/global/GenericTable";
+import { useRouter } from "next/navigation";
+
+interface AppointmentListProps {
+  petId: number;
+  token: string;
+}
+
+type AppointmentApiResponse = {
+  data: Appointment[];
+  total: number;
+  size: number;
+  prev: boolean;
+  next: boolean;
+  currentPage: number;
+  totalPages: number;
+};
+
+export default function AppointmentList({ petId, token }: AppointmentListProps) {
+  const router = useRouter();
+  const [appointments, setAppointments] = useState<Appointment[]>([]);
+  const [loadingAppointments, setLoadingAppointments] = useState(true);
+  const [errorAppointments, setErrorAppointments] = useState<string | null>(null);
+  const [pagination, setPagination] = useState<PaginationInfo>({
+    currentPage: 1,
+    totalPages: 1,
+    totalItems: 0,
+    pageSize: 100,
+  });
+
+  const estadoCitaEsp: Record<string, string> = {
+    COMPLETED: "Completado",
+    CANCELLED: "Cancelada",
+    PENDING: "Pendiente",
+    IN_PROGRESS: "En progreso",
+  };
+
+  const fetchAppointments = async (page = 1) => {
+    if (!petId || !token) return;
+    setLoadingAppointments(true);
+    try {
+      const data = await getAppointmentByPetId(petId, token, page, pagination.pageSize);
+      setAppointments(data);
+      setPagination(prev => ({
+        ...prev,
+        currentPage: page,
+        totalItems: data.length,
+      }));
+      setErrorAppointments(null);
+    } catch (error) {
+      setErrorAppointments("No se pudieron cargar las citas");
+      setAppointments([]);
+    } finally {
+      setLoadingAppointments(false);
+    }
+  };
+
+  useEffect(() => {
+    fetchAppointments(1);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [petId, token]);
+
+  const handlePageChange = (page: number) => {
+    fetchAppointments(page);
+  };
+
+  const appointmentColumns: Column<Appointment>[] = [
+    {
+      header: "Detalle",
+      accessor: "details",
+    },
+    {
+      header: "Fecha",
+      accessor: (a) => formatDate(a.designatedDate),
+    },
+    {
+      header: "Servicio",
+      accessor: "service",
+    },
+    {
+      header: "Empleados",
+      accessor: (a) => a.employees?.map(e => e.name).join(", ") || "Sin asignar",
+    },
+    {
+      header: "Estado",
+      accessor: (a) => estadoCitaEsp[a.status] || a.status,
+    },
+  ];
+
+  const appointmentActions: TableAction<Appointment>[] = [
+    {
+      icon: <Eye className="w-4 h-4" />,
+      onClick: (appointment: Appointment) => {
+        router.push(`/user-profile/appointment/${appointment.id}`);
+      },
+      label: "Ver detalle",
+    }
+  ];
+
+  if (errorAppointments) {
+    return <div className="text-red-500">{errorAppointments}</div>;
+  }
+
+  return (
+    <div>
+      <h2 className="text-2xl font-bold mb-3">Citas</h2>
+      <GenericTable<Appointment>
+        data={appointments}
+        columns={appointmentColumns}
+        actions={appointmentActions}
+        actionsTitle="Acciones"
+        pagination={pagination}
+        onPageChange={handlePageChange}
+        isLoading={loadingAppointments}
+        emptyMessage="Sin citas registradas"
+        className="mb-10"
+      />
+    </div>
+  );
+}
